BSP-167: SuccessfulTransformationResponse.warnings are array by default (this enables lombok getter), refactoring
