+++ build.gradle
@@ -21,7 +21,7 @@ plugins {
 apply plugin: 'org.springframework.boot'
 
 group 'uk.gov.hmcts.reform'
-version '0.0.14'
+version '0.0.15'
 
 dependencyUpdates.resolutionStrategy = {
     componentSelection { rules ->
