+++ src/main/java/uk/gov/hmcts/reform/bsp/common/model/update/output/SuccessfulUpdateResponse.java
@@ -3,7 +3,7 @@ package uk.gov.hmcts.reform.bsp.common.model.update.output;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
 import lombok.Data;
-import uk.gov.hmcts.reform.bsp.common.model.transformation.output.CaseUpdateDetails;
+import uk.gov.hmcts.reform.bsp.common.model.update.in.CaseUpdateDetails;
 
 import java.util.List;
